Implementing KS Test and Graph Analysis for Pathway Evaluation
This script introduces functions to analyze the dependency between the features of nodes in a given biological pathway and a random Barabasi-Albert graph. The Kolmogorov-Smirnov (KS) test is used to compare the distributions of the features in both graphs.

Key features of this implementation include:

run_KS_test() A function that computes the KS test statistic and p-value for two input feature lists.
callPathway_tomeasureNode_vitality() A function that reads a pathway graph from a GEXF file and returns it as a NetworkX graph.
evaluateNodesImportance_byFeatures() A function that computes various centrality metrics for nodes in the input graph, with optional exclusion of specific features.
buildFeature_analyseMatrix_wMCS_Grand() [which is the main function of script as higher in the hierarchy] A function that compares the biological pathway graph with a random Barabasi-Albert graph through multiple iterations, computes the KS test for each iteration, and returns the average KS coefficients for each feature.

Implementation allows for better understanding of the significance differences between the features of nodes in a biological pathway and a random graph to tune scoring function as finding coefficients.

This script is planned to use One-Time analyses to build all Pathways' scoring function (Coefficients) to save database retrieved from ConsensusPath.